feat(backend): change total from int to float64 for decimal precision
- Update Order model to use float64 for total field
- Modify pricing service to calculate totals with decimal precision
- Update repository to handle NUMERIC(10,2) database type
- Add user_id to order model and repository methods
- Update config to include JWT_SECRET environment variable

+++ publish-order-service/internal/service/pricing.go
@@ -3,7 +3,7 @@ package service
 import "github.com/icl00ud/publish-order-service/internal/model"
 
 type PricingCalculator interface {
-	Calculate(items []model.CartItem) int
+	Calculate(items []model.CartItem) float64
 }
 
 type defaultPricing struct{}
@@ -12,10 +12,10 @@ func NewPricingCalculator() PricingCalculator {
 	return defaultPricing{}
 }
 
-func (d defaultPricing) Calculate(items []model.CartItem) int {
-	sum := 0
+func (d defaultPricing) Calculate(items []model.CartItem) float64 {
+	sum := 0.0
 	for _, it := range items {
-		sum += int(it.Price * float64(it.Quantity))
+		sum += it.Price * float64(it.Quantity)
 	}
 	return sum
 }
